FileStorageService/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
python-dotenv
minio
python-multipart
